change deletions from deleting "branches" to removing references
previously broom was using the go-git method `DeleteBranches` which
deletes branches that are present in `.git/config`. This was causing
unexpected behavior since branches are only in `.git/config` if there is
a remote referfence to track with the local branch. This commit removes
the use of `DeleteBranches` in favor of `RemoveReferences` which instead
deletes the specific reference from the git store.